requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
groq>=0.9.0
tavily-python>=0.3.0
streamlit>=1.35.0
python-dotenv>=1.0.0
